Refactor http i13n tests
